vcmi: move CBonusProxy and friends to new file
To decouple HeroBonus.h more
